@@ -1,5 +1,5 @@
 name: aiimscycle
-description: "Simplify staff transportation with our user-friendly cycle rental service. Staff can quickly browse and book cycles through the app, promoting eco-friendly commuting within the AIIMS campus."
+description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
+  sdk: ^3.5.4
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,7 +34,7 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.6
+  cupertino_icons: ^1.0.8
   permission_handler: ^11.3.0
   dio: ^5.4.1
   flutter_bloc: ^8.1.4
@@ -43,7 +43,7 @@ dependencies:
   flutter_spinkit: ^5.2.0
   crypto: ^3.0.3
   fluttertoast: ^8.2.4
-  flutter_camera_qrcode_scanner: ^2.0.0
+#  flutter_camera_qrcode_scanner: ^2.0.0
   flutter_screenutil: ^5.9.0
   logger: ^2.2.0
   flutter_dotenv: ^5.1.0
@@ -59,9 +59,6 @@ dependencies:
   simple_barcode_scanner: ^0.1.1
   cached_network_image: ^3.3.1
   geolocator: ^11.0.0
-#  geolocator_android: 4.5.5
-
-
 
 dev_dependencies:
   flutter_test:
@@ -72,7 +69,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^3.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -93,10 +90,10 @@ flutter:
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware
+  # https://flutter.dev/to/resolution-aware-images
 
   # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
+  # https://flutter.dev/to/asset-from-package
 
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
@@ -116,4 +113,4 @@ flutter:
   #         weight: 700
   #
   # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
+  # see https://flutter.dev/to/font-from-package
